LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/bulk64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
FAIL #2 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "bulk64.wast:2"
FAIL #3 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "bulk64.wast:7"
FAIL #4 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "bulk64.wast:45"
FAIL #5 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "bulk64.wast:117"
FAIL #6 Test that WebAssembly compilation succeeds assert_equals: expected (boolean) false but got (string) "bulk64.wast:161"
PASS #7 Reinitialize the default imports
FAIL #8 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 16: Memory64 is not enabled at {loc} expected true got false
FAIL #9 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:7:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #10 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 43: Memory64 is not enabled at {loc} expected true got false
FAIL #11 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:13:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #12 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:16:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #13 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:19:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #14 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:22:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #15 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:25:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #16 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:28:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #17 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:31:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #18 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:34:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #19 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:37:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #20 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:40:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #21 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:43:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #22 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:46:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #23 Test that a WebAssembly code traps assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:244:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:49:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #24 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 43: Memory64 is not enabled at {loc} expected true got false
FAIL #25 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:55:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #26 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:58:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #27 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:61:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #28 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:64:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #29 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:67:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #30 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:70:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #31 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:73:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #32 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:76:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #33 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:79:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #34 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:82:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #35 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:85:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #36 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:88:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #37 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:91:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #38 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:94:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #39 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:97:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #40 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:100:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #41 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:103:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #42 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:106:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #43 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:109:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #44 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:112:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #45 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:115:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #46 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:118:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #47 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:121:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #48 Test that a WebAssembly code traps assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:244:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:124:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #49 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:127:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #50 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:130:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #51 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:133:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #52 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:136:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #53 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:139:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #54 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:142:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #55 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:145:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #56 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:148:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #57 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:151:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #58 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:154:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #59 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:157:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #60 Test that a WebAssembly code traps assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:244:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:160:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #61 Test that a WebAssembly code traps assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:244:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:163:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #62 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 43: Memory64 is not enabled at {loc} expected true got false
FAIL #63 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:169:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #64 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:172:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #65 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:175:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #66 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:178:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #67 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:181:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #68 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:184:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #69 Test that a WebAssembly code traps assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:244:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:187:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #70 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:190:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #71 Test that a WebAssembly code returns a specific result assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:267:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:193:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #72 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:196:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #73 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:199:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #74 Test that a WebAssembly code traps assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:244:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:202:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #75 Test that a WebAssembly code traps assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:244:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:205:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #76 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:208:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #77 Test that WebAssembly compilation succeeds assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 37: Memory64 is not enabled at {loc} expected true got false
FAIL #78 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed CompileError: WebAssembly.Module doesn't parse at byte 0: expected a module of at least 8 bytes instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:183:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:214:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #79 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:217:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #80 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:220:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #81 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:223:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #82 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:223:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:226:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
